+++ sources/FtpsServerWindows/Controls/NumericUpDown.xaml.cs
@@ -4,7 +4,7 @@ using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Input;
 
-namespace FtpsServerApp.Controls
+namespace FtpsServerWindows.Controls
 {
     public partial class NumericUpDown : UserControl
     {