+++ sources/FtpsServerWindows/Controls/ServerConfigurationControl.xaml.cs
@@ -1,11 +1,11 @@
 using FtpsServerApp.Helpers;
-using FtpsServerApp.Models;
+using FtpsServerWindows.Models;
 using Microsoft.Win32;
 using System;
 using System.Windows;
 using System.Windows.Controls;
 
-namespace FtpsServerApp.Controls
+namespace FtpsServerWindows.Controls
 {
     public partial class ServerConfigurationControl : UserControl
     {